Move TimeModels out of sync
