+++ shuttle/src/time.rs
@@ -1,0 +1,4 @@
+//! Timing primitives that allow Shuttle tests to interact with wall-clock time
+//! (`Instant`, `Duration`, `Timeout`, etc.) in a deterministic manner.
+
+pub use shuttle_engine::time::*;
