+++ shuttle/tests/basic/pct.rs
@@ -1,6 +1,6 @@
 use shuttle::scheduler::PctScheduler;
-use shuttle::sync::time::Duration;
 use shuttle::sync::Mutex;
+use shuttle::time::Duration;
 use shuttle::{check_pct, check_random, thread, Config, MaxSteps, Runner};
 use std::sync::atomic::AtomicUsize;
 use std::sync::atomic::Ordering;
